Fix configure of Refund action
